Changes done

+++ 00_simple/05_areatri.c
@@ -1,10 +1,11 @@
 #include<stdio.h>
+#include<math.h>
 
 int main()
 {
-	int a,b,c,s,area;
+	float a,b,c,s,area;
 	printf("Enter the lengths of sides\n");
-	scanf("%d %d %d", &a, &b, &c);
+	scanf("%f %f %f", &a, &b, &c);
 	s = a+b+c;
 	area = sqrt(s*(s-a)*(s-b)*(s-c));
 	printf("The area of the triangle is %d\n\n", area);
